requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]>=3.1.0
pydantic>=2.9.0
pydantic[email]>=2.9.0
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
passlib[argon2]==1.7.4
argon2-cffi>=23.1.0